pyproject.toml
[project]
name = "rpcstream"
version = "0.1.6"
description = "High-throughput blockchain ingestion engine for realtime streaming and historical backfill."
authors = [
  { name = "Chainlake" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "aiohttp>=3.10.0",
  "aiokafka>=0.11.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "orjson>=3.10.0",
  "uvloop>=0.19.0",
  "tenacity>=9.0.0",
  "structlog>=24.4.0",
  "pyyaml>=6.0.2",
  "opentelemetry-api>=1.17.0",
  "opentelemetry-sdk>=1.17.0",
  "opentelemetry-exporter-otlp>=1.17.0",
  "confluent-kafka>=2.14.0",
  "python-dotenv>=1.2.2",
  "pipreqs>=0.5.0",
  "pip-check-reqs>=2.5.6",
  "objgraph>=3.6.2",
  "authlib>=1.7.0",
  "httpx>=0.28.1",
  "cachetools>=7.0.6",
  "pytest>=9.0.3",
]

[project.optional-dependencies]

dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0"
]

[project.scripts]
rpcstream = "rpcstream.main:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
